requirements.txt
# Dipendenze del progetto
openai~=1.61.0
requests~=2.32.3
python-dotenv~=1.0.1
json5~=0.10.0